requirements.txt
certifi==2021.10.8         
charset-normalizer==2.0.12 
dnspython==2.2.1           
idna==3.3
maxminddb==2.2.0
numpy==1.19.5              
pandas==1.1.5              
pip==21.3.1                
python-dateutil==2.8.2     
pytz==2021.3               
requests==2.27.1           
setuptools==39.2.0         
simplejson==3.17.6         
six==1.16.0
texttable==1.6.4
urllib3==1.26.8   